Chat: split intelligence loop phase and fallback helpers
